refactor(tests): reorganize test files according to new architecture
feat(tests): add comprehensive test architecture documentation
docs: remove outdated CHANGELOG.md file

Move test files to their new locations under src/__tests__ following the standardized test architecture. Add detailed TEST_ARCHITECTURE.md documentation. Remove deprecated CHANGELOG.md as version history is now managed through git tags.
