+++ src/modules/__tests__/llm-provider-integration.test.ts
