+++ src/modules/__tests__/openai-embedding-provider.test.ts
