+++ src/persistence/__tests__/sqlite-graph-validation.test.ts
